refactor: renamed root file of each server to be server.js and installed modules on auth-server
