+++ auth-server/package.json
@@ -12,8 +12,10 @@
     "binary-parser": "^1.7.0",
     "common-errors": "^1.2.0",
     "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "helmet": "^4.4.1",
     "jsonwebtoken": "^8.5.1",
     "minimist": "^1.2.5",
     "mongoose": "^5.12.2"
-  },
+  }
 }
