+++ world-server/server.js
@@ -1,5 +1,5 @@
 require('dotenv').config()
 
-const { initialize } = require('./src/server')
+const { initialize } = require('./src/initializer')
 
 initialize({ port: process.env.PORT, host: process.env.HOST })
